made the code work, added docs-related content
