+++ .github/workflows/ghcr_sqlite_datacleaning.yml
@@ -1,0 +1,33 @@
+name: Build and Push SQLite3 Data Cleaning Docker Image
+
+on:
+  pull_request:
+  create:
+    tags:
+      - v*
+
+jobs:
+  push-github:
+    name: Push SQLite3 Data Cleaning examples to GHCR
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: cookbook
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: "0"
+      - name: Push Docker Image to Github Registry
+        uses: whoan/docker-build-with-cache-action@v5
+        with:
+          # https://docs.github.com/en/packages/learn-github-packages/publishing-a-package
+          username: "${{ secrets.FLYTE_BOT_USERNAME }}"
+          password: "${{ secrets.FLYTE_BOT_PAT }}"
+          image_name: ${{ github.repository_owner }}/flytecookbook
+          image_tag: sqlite_datacleaning-latest,sqlite_datacleaning-${{ github.sha }},sqlite_datacleaning-${{ github.event.ref }}
+          push_git_tag: ${{ github.event_name != 'pull_request' }}
+          push_image_and_stages: ${{ github.event_name != 'pull_request' }}
+          registry: ghcr.io
+          build_extra_args: "--compress=true --build-arg=tag=ghcr.io/${{ github.repository_owner }}/flytecookbook:sqlite_datacleaning-${{ github.sha }}"
+          context: ./cookbook/case_studies/feature_engineering/
+          dockerfile: sqlite_datacleaning/Dockerfile
