+++ cookbook/case_studies/feature_engineering/sqlite_datacleaning/datacleaning_tasks.py
@@ -1,3 +1,12 @@
+"""
+Data Cleaning Tasks
+-------------------------
+
+We'll define the relevant feature engineering tasks to clean up the SQLite3 data.
+"""
+
+# %%
+# Firstly, let's import the required libraries.
 import numpy as np
 import pandas as pd
 from flytekit import task
@@ -5,7 +14,8 @@ from numpy.core.fromnumeric import sort
 from sklearn.feature_selection import SelectKBest, f_classif
 from sklearn.impute import SimpleImputer
 
-
+# %%
+# Next, we define a ``mean_median_imputer`` task to fill in the missing values of the dataset, for which we use `SimpleImputer <https://scikit-learn.org/stable/modules/generated/sklearn.impute.SimpleImputer.html>`__ class picked from the ``scikit-learn`` library.
 @task
 def mean_median_imputer(
     dataframe: pd.DataFrame,
@@ -23,6 +33,12 @@ def mean_median_imputer(
 
     return dataframe
 
+# %%
+# This task returns the filled-in dataframe.
+
+# %%
+# Let's define one other task called ``univariate_selection`` that does feature selection. 
+# The `SelectKBest <https://scikit-learn.org/stable/modules/generated/sklearn.feature_selection.SelectKBest.html#sklearn.feature_selection.SelectKBest>`__ method removes all but the highest scoring features.
 
 @task
 def univariate_selection(
@@ -37,3 +53,6 @@ def univariate_selection(
     column_names = map(dataframe.columns.__getitem__, indices)
     features = fit.transform(X)
     return pd.DataFrame(features, columns=column_names)
+
+# %%
+# This task returns a dataframe with the specified number of columns.
