+++ cookbook/docs/feature_engineering.rst
@@ -2,6 +2,19 @@
 Feature Engineering
 ###################
 
-.. NOTE::
+.. panels::
+    :header: text-center
 
-    Coming soon 🛠
+    .. link-button:: auto/case_studies/feature_engineering/sqlite_datacleaning/index
+       :type: ref
+       :text: Data Cleaning
+       :classes: btn-block stretched-link
+    ^^^^^^^^^^^^
+    Perform data imputation and univariate feature selection on SQLite3 dataset
+
+.. toctree::
+    :maxdepth: -1
+    :caption: Contents
+    :hidden:
+
+    auto/case_studies/feature_engineering/sqlite_datacleaning/index
