+++ cookbook/docs/ml_training.rst
@@ -2,13 +2,33 @@
 ML Training
 ################
 
+.. panels::
+    :header: text-center
+
+    .. link-button:: auto/case_studies/ml_training/pima_diabetes/index
+       :type: ref
+       :text: Diabetes Classification
+       :classes: btn-block stretched-link
+    ^^^^^^^^^^^^
+    Train an XGBoost model on the Pima Indians Diabetes Dataset
+
+    ---
+
+    .. link-button:: auto/case_studies/ml_training/house_price_prediction/index
+       :type: ref
+       :text: House Price Regression
+       :classes: btn-block stretched-link
+    ^^^^^^^^^^^^
+    Use dynamic workflows to train a multiregion house price prediction model.
+
+
 .. toctree::
     :maxdepth: -1
     :caption: Contents
+    :hidden:
 
     auto/case_studies/ml_training/pima_diabetes/index
     auto/case_studies/ml_training/house_price_prediction/index
 
-.. admonition:: Coming Soon!
 
-    Data Parallel Training, Distributed Training, and Single Node Training 
+.. TODO: write tutorials for data parallel training, distributed training, and single node training
